laywer_profile/requirements.txt
Scrapy-2.8.0
requests-2.28.2
csvwriter-0.2.2
Selector-0.10.1
parse 1.19.0